#1220 move onmounted logic to App.vue
